fix: dynamically sets stepnav from default edit view
